+++ app/imports/ui/components/includes/file-uploader/file-uploader.js
@@ -2,7 +2,7 @@ import { Template } from 'meteor/templating';
 import { Random } from 'meteor/random';
 import { ReactiveArray } from 'meteor/manuel:reactivearray';
 
-import UploadService from '/imports/ui/components/uploads/UploadService';
+import UploadService from '/imports/ui/utils/uploads/UploadService';
 
 
 Template.FileUploader.viewmodel({
