+++ app/imports/ui/components/index.js
@@ -17,4 +17,3 @@ import './standard-subcards';
 import './discussion';
 import './notifications';
 import './problems';
-import './uploads';
