+++ app/imports/ui/utils/index.js
@@ -1,0 +1,1 @@
+import './uploads';
